Move tasks out of commands folder into their own folder.
