+++ lib/digitalocean-deploy-plugin.js
@@ -3,10 +3,10 @@ var SSHClient = require('ssh2').Client;
 var SCPClient = require('scp2');
 var path = require('path');
 var RSVP = require('rsvp');
-var NPMInstallTask = require('./commands/npm-install');
-var RemoteCommandTask = require('./commands/remote-command');
-var UploadAppTask = require('./commands/upload-app');
-var UploadFileTask = require('./commands/upload-file');
+var NPMInstallTask = require('./tasks/npm-install');
+var RemoteCommandTask = require('./tasks/remote-command');
+var UploadAppTask = require('./tasks/upload-app');
+var UploadFileTask = require('./tasks/upload-file');
 
 const CONFIG_ENV_MAPPING = {
   FASTBOOT_DO_APPLICATION: 'dropletName',
