+++ lib/tasks/npm-install.js
