+++ lib/tasks/upload-app.js
