+++ lib/tasks/upload-file.js
